+++ CSS535_Project_Upscale/bitmap.cpp
@@ -15,12 +15,12 @@ unsigned char Bitmap::padSize()
 	return Bitmap::padSize(this->width);
 }
 
-unsigned int Bitmap::imageDataSize(unsigned int width, unsigned int height)
+unsigned int Bitmap::imageDataSize(unsigned short width, unsigned short height)
 {
 	return width * height * 3 + height * Bitmap::padSize(width);
 }
 
-unsigned char Bitmap::padSize(unsigned int width)
+unsigned char Bitmap::padSize(unsigned short width)
 {
 	return  (4 - (width * 3) % 4) % 4;
 }
